fix: move getNewPermissions to utils
