+++ src/infrastructure/commands/MigrateCommand.ts
@@ -111,5 +111,4 @@ export class MigrateCommand {
             module: permissionModule,
         });
     }
-
 }
